NDEV-2183: Implement get_sync_status for eth_syncing
